# refactor: use parallelizable scheduler for explore + entity