--- a/hypertrace-graphql-entity-schema/src/main/java/org/hypertrace/graphql/entity/dao/GatewayBaselineDao.java
+++ b/hypertrace-graphql-entity-schema/src/main/java/org/hypertrace/graphql/entity/dao/GatewayBaselineDao.java
@@ -5,6 +5,7 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.hypertrace.core.graphql.common.utils.CollectorUtils.flatten;
 
 import io.grpc.CallCredentials;
+import io.reactivex.rxjava3.core.Scheduler;
 import io.reactivex.rxjava3.core.Single;
 import java.util.Collection;
 import java.util.Set;
@@ -13,9 +14,10 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 import org.hypertrace.core.graphql.common.utils.Converter;
 import org.hypertrace.core.graphql.context.GraphQlRequestContext;
+import org.hypertrace.core.graphql.rx.BoundedIoScheduler;
 import org.hypertrace.core.graphql.spi.config.GraphQlServiceConfig;
-import org.hypertrace.core.graphql.utils.grpc.GraphQlGrpcContextBuilder;
 import org.hypertrace.core.graphql.utils.grpc.GrpcChannelRegistry;
+import org.hypertrace.core.graphql.utils.grpc.GrpcContextBuilder;
 import org.hypertrace.gateway.service.GatewayServiceGrpc;
 import org.hypertrace.gateway.service.v1.baseline.BaselineEntitiesRequest;
 import org.hypertrace.gateway.service.v1.baseline.BaselineEntitiesResponse;
@@ -34,9 +36,9 @@ import org.hypertrace.graphql.metric.request.MetricSeriesRequest;
 
 @Singleton
 class GatewayBaselineDao implements BaselineDao {
-  private static final int DEFAULT_DEADLINE_SEC = 10;
   private final GatewayServiceGrpc.GatewayServiceFutureStub gatewayServiceStub;
-  private final GraphQlGrpcContextBuilder grpcContextBuilder;
+  private final GrpcContextBuilder grpcContextBuilder;
+  private final Scheduler boundedIoScheduler;
   private final Converter<Collection<MetricAggregationRequest>, Set<Expression>>
       aggregationConverter;
   private final Converter<Collection<MetricSeriesRequest>, Set<TimeAggregation>> seriesConverter;
@@ -47,10 +49,12 @@ class GatewayBaselineDao implements BaselineDao {
       GrpcChannelRegistry grpcChannelRegistry,
       GraphQlServiceConfig serviceConfig,
       CallCredentials credentials,
-      GraphQlGrpcContextBuilder grpcContextBuilder,
+      GrpcContextBuilder grpcContextBuilder,
+      @BoundedIoScheduler Scheduler boundedIoScheduler,
       Converter<Collection<MetricAggregationRequest>, Set<Expression>> aggregationConverter,
       Converter<Collection<MetricSeriesRequest>, Set<TimeAggregation>> seriesConverter) {
     this.grpcContextBuilder = grpcContextBuilder;
+    this.boundedIoScheduler = boundedIoScheduler;
     this.gatewayServiceStub =
         GatewayServiceGrpc.newFutureStub(
                 grpcChannelRegistry.forAddress(
@@ -68,6 +72,7 @@ class GatewayBaselineDao implements BaselineDao {
       EntitiesResponse entitiesResponse,
       EntityRequest request) {
     return this.buildRequest(entitiesRequest, entitiesResponse, request)
+        .subscribeOn(this.boundedIoScheduler)
         .flatMap(baselineEntitiesRequest -> makeRequest(context, baselineEntitiesRequest));
   }
 
@@ -126,7 +131,7 @@ class GatewayBaselineDao implements BaselineDao {
     return Single.fromFuture(
         this.grpcContextBuilder
             .build(context)
-            .callInContext(
+            .call(
                 () ->
                     this.gatewayServiceStub
                         .withDeadlineAfter(