--- a/hypertrace-graphql-entity-schema/src/main/java/org/hypertrace/graphql/entity/dao/GatewayServiceEntityDao.java
+++ b/hypertrace-graphql-entity-schema/src/main/java/org/hypertrace/graphql/entity/dao/GatewayServiceEntityDao.java
@@ -3,13 +3,15 @@ package org.hypertrace.graphql.entity.dao;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 import io.grpc.CallCredentials;
+import io.reactivex.rxjava3.core.Scheduler;
 import io.reactivex.rxjava3.core.Single;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import org.hypertrace.core.graphql.context.GraphQlRequestContext;
+import org.hypertrace.core.graphql.rx.BoundedIoScheduler;
 import org.hypertrace.core.graphql.spi.config.GraphQlServiceConfig;
-import org.hypertrace.core.graphql.utils.grpc.GraphQlGrpcContextBuilder;
 import org.hypertrace.core.graphql.utils.grpc.GrpcChannelRegistry;
+import org.hypertrace.core.graphql.utils.grpc.GrpcContextBuilder;
 import org.hypertrace.gateway.service.GatewayServiceGrpc;
 import org.hypertrace.gateway.service.GatewayServiceGrpc.GatewayServiceFutureStub;
 import org.hypertrace.gateway.service.v1.entity.EntitiesRequest;
@@ -21,26 +23,29 @@ import org.hypertrace.graphql.entity.schema.EntityResultSet;
 @Singleton
 class GatewayServiceEntityDao implements EntityDao {
   private final GatewayServiceFutureStub gatewayServiceStub;
-  private final GraphQlGrpcContextBuilder grpcContextBuilder;
+  private final GrpcContextBuilder grpcContextBuilder;
   private final GatewayServiceEntityRequestBuilder requestBuilder;
   private final GatewayServiceEntityConverter entityConverter;
   private final BaselineDao baselineDao;
   private final GraphQlServiceConfig serviceConfig;
+  private final Scheduler boundedIoScheduler;
 
   @Inject
   GatewayServiceEntityDao(
       GraphQlServiceConfig serviceConfig,
       CallCredentials credentials,
-      GraphQlGrpcContextBuilder grpcContextBuilder,
+      GrpcContextBuilder grpcContextBuilder,
       GrpcChannelRegistry grpcChannelRegistry,
       GatewayServiceEntityRequestBuilder requestBuilder,
       GatewayServiceEntityConverter entityConverter,
-      BaselineDao baselineDao) {
+      BaselineDao baselineDao,
+      @BoundedIoScheduler Scheduler boundedIoScheduler) {
     this.grpcContextBuilder = grpcContextBuilder;
     this.requestBuilder = requestBuilder;
     this.entityConverter = entityConverter;
     this.baselineDao = baselineDao;
     this.serviceConfig = serviceConfig;
+    this.boundedIoScheduler = boundedIoScheduler;
 
     this.gatewayServiceStub =
         GatewayServiceGrpc.newFutureStub(
@@ -53,6 +58,7 @@ class GatewayServiceEntityDao implements EntityDao {
   public Single<EntityResultSet> getEntities(EntityRequest request) {
     return this.requestBuilder
         .buildRequest(request)
+        .subscribeOn(this.boundedIoScheduler)
         .flatMap(
             serverRequest ->
                 this.makeRequest(request.resultSetRequest().context(), serverRequest)
@@ -75,7 +81,7 @@ class GatewayServiceEntityDao implements EntityDao {
     return Single.fromFuture(
         this.grpcContextBuilder
             .build(context)
-            .callInContext(
+            .call(
                 () ->
                     this.gatewayServiceStub
                         .withDeadlineAfter(