--- a/hypertrace-graphql-explorer-schema/build.gradle.kts
+++ b/hypertrace-graphql-explorer-schema/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
   implementation("org.hypertrace.core.graphql:hypertrace-core-graphql-attribute-store")
   implementation("org.hypertrace.core.graphql:hypertrace-core-graphql-deserialization")
   implementation("org.hypertrace.core.graphql:hypertrace-core-graphql-schema-utils")
+  implementation("org.hypertrace.core.graphql:hypertrace-core-graphql-rx-utils")
 
   testImplementation("org.junit.jupiter:junit-jupiter")
   testImplementation("org.mockito:mockito-core")